+++ Kernel/Modules/AdminDynamicFieldDropdown.pm
@@ -101,6 +101,8 @@ sub _Add {
         %Param,
         %GetParam,
         Mode           => 'Add',
+        BreadcrumbText => $LayoutObject->{LanguageObject}
+            ->Translate( 'Add %s field', $LayoutObject->{LanguageObject}->Translate($FieldTypeName) ),
         ObjectTypeName => $ObjectTypeName,
         FieldTypeName  => $FieldTypeName,
     );
@@ -338,6 +340,8 @@ sub _Change {
         %Config,
         ID             => $FieldID,
         Mode           => 'Change',
+        BreadcrumbText => $LayoutObject->{LanguageObject}
+            ->Translate( 'Change %s field', $LayoutObject->{LanguageObject}->Translate($FieldTypeName) ),
         ObjectTypeName => $ObjectTypeName,
         FieldTypeName  => $FieldTypeName,
     );